Remove ezlvm and ffs
These components to depend on the smapiv3 component (to provide
xapi-storage-plugins) but also on the other ocaml packages. Therefore it's not
possible to have smapiv3 depending on dom0 because it depends on ocaml.

For now, let's not build them to break this circular dependency. This is fine
since they are not used in the product.